requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.2
django-cors-headers==3.7.0
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
fake-headers==1.0.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
indicator==1.0.5
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mpl-finance==0.10.1
mplfinance==0.12.6a3
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
nsepy==0.8
numpy==1.19.0
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pandas-datareader==0.8.1
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
plotly==4.8.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.13.0
PyQt5-sip==12.8.0
pyqtgraph==0.11.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
retrying==1.3.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
xlrd==1.2.0
